fix: html 외 다른 확장자 미지원 오류로 viewController 제거
