+++ tomcat/src/main/java/com/techcourse/controller/FrontController.java
@@ -1,16 +1,25 @@
 package com.techcourse.controller;
 
+import java.io.File;
 import java.io.IOException;
+import java.net.URL;
+import java.nio.file.Files;
+import java.nio.file.Path;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Objects;
 
 import org.apache.coyote.http11.HttpRequest;
 import org.apache.coyote.http11.HttpResponse;
+import org.apache.coyote.http11.HttpStatus;
+import org.apache.coyote.http11.MimeType;
+import org.apache.coyote.http11.ResponseBody;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import com.techcourse.exception.InvalidResourceException;
 import com.techcourse.exception.UnsupportedMethodException;
+import com.techcourse.util.FileExtension;
 
 public class FrontController extends Controller {
     private static final FrontController instance = new FrontController();
@@ -27,9 +36,6 @@ public class FrontController extends Controller {
     }
 
     private void initHandlerMappings() {
-        handlerMappings.put("/", ViewController.getInstance());
-        handlerMappings.put("/hello", ViewController.getInstance());
-        handlerMappings.put("/index", ViewController.getInstance());
         handlerMappings.put("/login", LoginController.getInstance());
         handlerMappings.put("/register", RegisterController.getInstance());
     }
@@ -37,6 +43,21 @@ public class FrontController extends Controller {
     @Override
     public HttpResponse handle(HttpRequest request) throws IOException {
         Controller handler = getHandler(request.getURI());
+        String fileName = getFileName(request.getURI());
+        if (Objects.isNull(handler) && FileExtension.isFileExtension(fileName)) {
+            try {
+                HttpResponse response = new HttpResponse();
+                ResponseBody responseBody = new ResponseBody(readResource(fileName));
+                response.setStatus(HttpStatus.OK);
+                response.setContentType(MimeType.getMimeType(fileName));
+                response.setBody(responseBody);
+                return response;
+            } catch (InvalidResourceException e) {
+                log.error("Error processing request for endpoint: {}", request.getURI());
+
+                handler = NotFoundController.getInstance();
+            }
+        }
         if (Objects.isNull(handler)) {
             log.error("Error processing request for endpoint: {}", request.getURI());
 
@@ -56,6 +77,32 @@ public class FrontController extends Controller {
 
     private Controller getHandler(String uri) {
         return handlerMappings.get(uri);
+    }
+
+    private String getFileName(String endpoint) {
+        int index = endpoint.indexOf("?");
+        String path = endpoint;
+        if (index != -1) {
+            path = path.substring(0, index);
+        }
+        String fileName = path.substring(1);
+        if (fileName.isEmpty()) {
+            fileName = "hello.html";
+        }
+        return fileName;
+    }
+
+    private String readResource(String fileName) throws IOException {
+        URL resource = findResource(fileName);
+        if (Objects.isNull(resource)) {
+            throw new InvalidResourceException("Cannot find resource with name: " + fileName);
+        }
+        Path path = new File(resource.getFile()).toPath();
+        return Files.readString(path);
+    }
+
+    private URL findResource(String fileName) {
+        return getClass().getClassLoader().getResource("static/" + fileName);
     }
 
     @Override
