+++ tomcat/src/main/java/org/apache/coyote/http11/MimeType.java
@@ -39,8 +39,4 @@ public enum MimeType {
     public String getType() {
         return type;
     }
-
-    public String getExtension() {
-        return extension;
-    }
 }
